bump to sbt 1.7.2 and scala 2.13, get rid of slide system

+++ mainpage/src/main/scala/mainobject/Loader.scala
@@ -6,7 +6,7 @@ object Loader {
     items.head.apply()
 
     if (items.tail.nonEmpty) {
-      scala.scalajs.js.timers.setTimeout(rate) {
+      scala.scalajs.js.timers.setTimeout(rate.toDouble) {
         load(items.tail, rate)
       }
     }
